Remove unused function
